decompose GA module
